v7.0b
Light control is back! Testing willing.

Zendo will adjust the brightness of the lights in your room in an iPhone meditation. If this feature is useful, we can may drive it into the watch. Wanted to get this done as it is the final piece of the external driven experience.